Refactor plugin renamers to not be based on the plugin instance being the renamer
